tests/wpt/tests/screen-wake-lock/wakelock-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that screen-wake-lock is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissions-policy-features">
<link rel="help" href="https://w3c.github.io/screen-wake-lock/#dfn-wake-lock-feature">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('screen-wake-lock', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise screen-wake-lock.');
</script>
